Impedance Matcher
This circuit allows to connect a guitar to a line-in without loosing
audio quality. For example, it can be connected ti a PC or development
board and keep the original sound. This circuit is an improved version
of an old circuit I made. I increased 2M resistors to 10M and 10uF
electrolitic capacitor to 47uF. It has been also added a low-pass filter
to cut the high frequencies and reduce the bandwidth of the system.